+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Technical - On start, create a forestadmin-schema.json file that contains the apimap definition.
 
 ## RELEASE 2.14.6 - 2018-12-14
 ### Changed